% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docshop_Package.Documentation_function.R
\docType{package}
\name{docshop-package}
\alias{docshop}
\alias{docshop-package}
\title{docshop: Doctor Shopping Functions}
\description{
Personal Functions for Doctor Shopping Project
}
\details{
This package includes many functions created by BK for studying doctor shopping patterns in the U.S. It is not intended for public use. The documentation for functions in this package is incomplete.
}
\author{
\strong{Maintainer}: Brian Aronson \email{bdaronson@gmail.com}

}
