% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtHaversine.R
\name{dtHaversine}
\alias{dtHaversine}
\title{Calculate Haversine distance using data.table framework}
\usage{
dtHaversine(lat_from, lon_from, lat_to, lon_to, r = 6378137)
}
\description{
Calculate Haversine distance using data.table framework
}
