% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elapsed_quarter_from_origin.R
\name{elapsed_quarter_from_origin}
\alias{elapsed_quarter_from_origin}
\title{Determine how many quarters elapsed between time period(?)}
\usage{
elapsed_quarter_from_origin(yy, qq, reference_year = 2007,
  reference_quarter = 1)
}
\description{
Determine how many quarters elapsed between time period(?)
}
