% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ctype_from_file.R
\name{extract_ctype_from_file}
\alias{extract_ctype_from_file}
\title{Extracts community type from file}
\usage{
extract_ctype_from_file(infile)
}
\description{
Extracts community type from file
}
