% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datatype_from_filename.R
\name{extract_datatype_from_filename}
\alias{extract_datatype_from_filename}
\title{Extracts datatype from filename}
\usage{
extract_datatype_from_filename(outfile)
}
\description{
Extracts datatype from filename
}
