% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_direction_from_file.R
\name{extract_direction_from_file}
\alias{extract_direction_from_file}
\title{Extracts direction from file}
\usage{
extract_direction_from_file(outfile)
}
\description{
Extracts direction from file
}
