% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_from_filename.R
\name{extract_year_from_filename}
\alias{extract_year_from_filename}
\title{Extracts year from filename}
\usage{
extract_year_from_filename(outfile)
}
\description{
Extracts year from filename
}
