% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year_quarter_from_filename.R
\name{extract_year_quarter_from_filename}
\alias{extract_year_quarter_from_filename}
\title{Extract year and quarter from filename}
\usage{
extract_year_quarter_from_filename(outfile)
}
\description{
Extract year and quarter from filename
}
