% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_first_last_date_table.R
\name{generate_first_last_date_table}
\alias{generate_first_last_date_table}
\title{generate_first_last_date_table; unsure}
\usage{
generate_first_last_date_table(from_year = 2007, to_year = 2017,
  time_unit = "quarter")
}
\description{
generate_first_last_date_table; unsure
}
