% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_average.R
\name{geographic_average}
\alias{geographic_average}
\title{Calculates median latitudes and longitudes based on input}
\usage{
geographic_average(lat, lng, weight = NULL)
}
\description{
Calculates median latitudes and longitudes based on input
}
