% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_previous_year_quarter.R
\name{pick_previous_year_quarter}
\alias{pick_previous_year_quarter}
\title{Find quarter of prior time point based on desired lag}
\usage{
pick_previous_year_quarter(year, quarter, lag)
}
\description{
Find quarter of prior time point based on desired lag
}
