% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_halfyear_date.R
\name{reference_halfyear_date}
\alias{reference_halfyear_date}
\title{Find input date based on year and half-yeard data}
\usage{
reference_halfyear_date(outfile)
}
\description{
Find input date based on year and half-yeard data
}
