% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_quarter_date.R
\name{reference_quarter_date}
\alias{reference_quarter_date}
\title{Create dates from year/quarter info}
\usage{
reference_quarter_date(outfile)
}
\description{
Create dates from year/quarter info
}
