% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3list.files.R
\name{s3.list.files}
\alias{s3.list.files}
\title{s3.list.files}
\usage{
s3.list.files(x, full.names = F, max.files = 1000)
}
\description{
A version of R's list.files function compatible with s3. By default, this is set to return up to only 1000 file names, only only the short name (i.e. excluding file paths). Note: A key difference between this function and list.files is that this function will return all files within subdirectories of a folder.
}
