% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3read_any.R
\name{s3read_any}
\alias{s3read_any}
\title{Reads fst, csv, and gz objects from S3; may cause errors for csv.gz}
\usage{
s3read_any(awsfile, localdirec = NULL, localfile = TRUE, ...)
}
\arguments{
\item{localdirec}{A local directory that might have that same file}

\item{localfile}{Search for file locally before downloading it from AWS}

\item{awfile}{The path and file you want to read from aws}
}
\description{
Reads fst, csv, and gz objects from S3; may cause errors for csv.gz
}
