% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3write_fst.R
\name{s3write_fst}
\alias{s3write_fst}
\title{Save object to Amazon S3 in .fst format}
\usage{
s3write_fst(data, object, compress = 100)
}
\arguments{
\item{data}{Data to save}

\item{object}{File name to save as}

\item{compress}{Degree of file compression}
}
\description{
Writes objects to Amazon S3 in fst format.
}
\examples{
s3write_fst(
  list_mdata,
  object = paste0("s3://iuni-doctorshopping/derived_R/edgelist/","edgelist_all",".fst"),
  compress=100
)
}
\keyword{S3}
\keyword{write}
