% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_generate.R
\name{tmp_generate}
\alias{tmp_generate}
\title{Generate tempfile name}
\usage{
tmp_generate(ext = ".fst")
}
\arguments{
\item{ext}{File extension for the temp file.}
}
\description{
Creates a tempfile name for use when saving objects to S3 (i.e. to prevent data from being permanently saved to the computer during intermediary steps).
}
